LayoutTests/platform/ios-sim/svg/text/caret-in-svg-text-expected.txt

CONSOLE MESSAGE: line 23: ReferenceError: Can't find variable: textInputController
Sheriff Woody
שדגש
